--- src/components/Placeholder/Placeholder.stories.tsx
@@ -1,8 +1,0 @@
-import React, { ReactElement } from 'react';
-import Placeholder from '.';
-
-export default {
-    title: 'Placeholder',
-};
-
-export const Default = (): ReactElement => <Placeholder />;
