--- src/components/Placeholder/index.tsx
@@ -1,3 +1,0 @@
-import Placeholder from './Placeholder';
-
-export default Placeholder;
